+++ src/index.js
@@ -1,6 +1,9 @@
 import path from 'node:path/posix';
 
 import secretManager from '@google-cloud/secret-manager';
+import Debug from 'debug';
+
+const debug = Debug('aller:google-cloud-secret');
 
 export class ConcurrentSecretError extends Error {
   /**
@@ -17,10 +20,10 @@ export class ConcurrentSecretError extends Error {
 export class ConcurrentSecret {
   /**
    * @param {string} name secret resource name, e.g. `projects/1234/secrets/concurrent-test-secret`
-   * @param {import('google-gax').ClientOptions | import('@google-cloud/secret-manager').v1.SecretManagerServiceClient} [clientOptions]
-   * @param {number} [gracePeriodMs] lock grace period in milliseconds, continue if secret is locked beyond grace period
+   * @param {import('google-gax').ClientOptions | import('@google-cloud/secret-manager').v1.SecretManagerServiceClient} [clientOptions] Secret Manager client instance or the options for a new one
+   * @param {concurrentSecretOptions} [options] options
    */
-  constructor(name, clientOptions, gracePeriodMs) {
+  constructor(name, clientOptions, options) {
     this.name = name;
     this.latestVersionName = path.join(name, '/versions/latest');
     this.client =
@@ -37,17 +40,8 @@ export class ConcurrentSecret {
     /** @type {import('@google-cloud/secret-manager').protos.google.cloud.secretmanager.v1.ISecretVersion | undefined} */
     this.secretVersion = undefined;
 
-    /** @type {number} [gracePeriodMs] Lock grace period in milliseconds, continue if secret is locked beyond grace period */
-    this.gracePeriodMs = gracePeriodMs ?? 60000;
-  }
-  /**
-   * Prepare optimistic update
-   * @internal
-   */
-  _prepare() {
-    if (this.pendingSecret) return this.pendingSecret;
-    this.pendingSecret = this.client.getSecret({ name: this.name }).then(([secret]) => secret);
-    return this.pendingSecret;
+    /** @type {concurrentSecretOptions} [gracePeriodMs] Lock grace period in milliseconds, continue if secret is locked beyond grace period */
+    this.options = { gracePeriodMs: 60000, ...options };
   }
   /**
    * Get latest version
@@ -70,7 +64,7 @@ export class ConcurrentSecret {
    * Get latest version secret data
    */
   async getLatestData() {
-    const [data] = await this.client.accessSecretVersion({ name: this.latestVersionName });
+    const [data] = await this.client.accessSecretVersion({ name: this.latestVersionName }, this._getCallOptions());
     return data;
   }
   /**
@@ -87,22 +81,25 @@ export class ConcurrentSecret {
 
       const latestVersion = await this.getLatestVersion();
 
-      await this.client.addSecretVersion({ parent: secret.name, payload: { data: Buffer.from(secretData) } });
+      await this.client.addSecretVersion(
+        {
+          parent: secret.name,
+          payload: { data: Buffer.from(secretData) },
+        },
+        this._getCallOptions()
+      );
+
       if (latestVersion && latestVersion.state !== 'DESTROYED' && !latestVersion.scheduledDestroyTime) {
         await this.client.destroySecretVersion({ name: latestVersion.name });
+        debug('secret version %s destroyed', latestVersion.name);
       }
 
-      const [updatedSecret] = await this.client.updateSecret({
-        secret: {
-          name: secret.name,
-          etag: secret.etag,
-          annotations: {
-            ...secret.annotations,
-            updated_at: new Date().toISOString(),
-          },
-        },
-        updateMask: {
-          paths: ['annotations'],
+      const [updatedSecret] = await this._updateSecret({
+        name: secret.name,
+        etag: secret.etag,
+        annotations: {
+          ...secret.annotations,
+          updated_at: new Date().toISOString(),
         },
       });
 
@@ -129,28 +126,24 @@ export class ConcurrentSecret {
 
     // @ts-ignore
     if (!isNaN(lockedAt)) {
-      const gracePeriodEat = new Date(lockedAt.getTime() + this.gracePeriodMs);
-
+      const gracePeriodEat = new Date(lockedAt.getTime() + this.options.gracePeriodMs);
       if (now <= gracePeriodEat) {
         throw new ConcurrentSecretError(`Secret is updated by another process since ${lockedAt.toISOString()}`, 9);
       }
     }
 
-    const [lockedSecret] = await this.client.updateSecret({
-      secret: {
-        name: secret.name,
-        etag: secret.etag,
-        annotations: {
-          ...secret.annotations,
-          locked_at: now.toJSON(),
-        },
-      },
-      updateMask: {
-        paths: ['annotations'],
+    const [lockedSecret] = await this._updateSecret({
+      name: secret.name,
+      etag: secret.etag,
+      annotations: {
+        ...secret.annotations,
+        locked_at: now.toJSON(),
       },
     });
 
     this.secret = lockedSecret;
+
+    debug('secret %s locked with etag %s', lockedSecret.name, lockedSecret.etag);
 
     return lockedSecret;
   }
@@ -167,18 +160,59 @@ export class ConcurrentSecret {
     // @ts-ignore
     const { locked_at, ...annotations } = secret.annotations;
 
-    const [updatedSecret] = await this.client.updateSecret({
-      secret: {
-        name: secret.name,
-        etag: secret.etag,
-        annotations,
-      },
-      updateMask: {
-        paths: ['annotations'],
-      },
+    const [updatedSecret] = await this._updateSecret({
+      name: secret.name,
+      etag: secret.etag,
+      annotations,
     });
+
     this.secret = updatedSecret;
+
+    debug('secret %s released with etag %s', updatedSecret.name, updatedSecret.etag);
+  }
+  /**
+   * @internal Prepare optimistic update
+   */
+  _prepare() {
+    if (this.pendingSecret) return this.pendingSecret;
+
+    debug('preparing secret %s', this.name);
+
+    this.pendingSecret = this.client.getSecret({ name: this.name }, this._getCallOptions()).then(([secret]) => secret);
+    return this.pendingSecret;
+  }
+  /**
+   * @internal Get gax call options
+   * @param {import('@google-cloud/secret-manager').protos.google.cloud.secretmanager.v1.ISecret} secret
+   */
+  _updateSecret(secret) {
+    return this.client.updateSecret(
+      {
+        secret,
+        updateMask: {
+          paths: secret.annotations && ['annotations'],
+        },
+      },
+      this._getCallOptions()
+    );
+  }
+  /**
+   * @internal Get gax call options
+   * @returns {import('google-gax').CallOptions}
+   */
+  _getCallOptions() {
+    if (typeof this.options.callOptions === 'function') {
+      return this.options.callOptions();
+    }
+    // @ts-ignore
+    return { ...this.options.callOptions };
   }
 }
 
 export default ConcurrentSecret;
+
+/**
+ * @typedef {object} concurrentSecretOptions
+ * @property {number} [gracePeriodMs] lock grace period in milliseconds, continue if secret is locked beyond grace period, default is 60000ms
+ * @property {()=>import('google-gax').CallOptions|import('google-gax').CallOptions} [callOptions] optional function to pass other args to pass to each request, tracing for instance
+ */