+++ test/helpers/fake-server.js
@@ -2,7 +2,7 @@ import fs from 'node:fs';
 
 import { startServer as startFakeServer } from '../../src/fake-server/fake-secret-manager-server.js';
 
-export { reset, RpcCodes } from '../../src/fake-server/fake-secret-manager-server.js';
+export { reset, RpcCodes, getSecret } from '../../src/fake-server/fake-secret-manager-server.js';
 
 const cert = {
   private_key: fs.readFileSync('./tmp/mkcert/dev-key.pem'),